Crate syntastica_core

syntastica-core

Shared types and traits used by other syntastica crates.

See the project overview for more information.

Features

runtime-c (enabled by default) — Use the standard tree-sitter C runtime. See syntastica’s WebAssembly support for more information.

runtime-c2rust — Use the pure Rust tree-sitter runtime. See syntastica’s WebAssembly support for more information.

serde — Provide implementations of serde’s Serialize and Deserialize traits for Theme and Style

docs — Enable this when building the docs

Modules

language_set

Defines the LanguageSet trait and some related types.

style

Defines the Style and Color types used by ResolvedThemes.

theme

Defines items related to theming the output.

Macros

theme

Convenience macro for constructing new Themes.

Enums

Error

The main error type.

Type Aliases

Result

The main result type.