Enum syntastica_core::language_set::FileType

pub enum FileType {

    Text,
    A2ps,
    A65,
    Aap,
    Abap,
    Abaqus,
    Abc,
    Abel,
    Acedb,
    Ada,
    Ahdl,
    Aidl,
    AlsaConf,
    Aml,
    Ampl,
    Ant,
    Apache,
    ApacheStyle,
    AppleScript,
    AptConf,
    Arch,
    Arduino,
    Art,
    AsciiDoc,
    Asm,
    Asn,
    AspPerl,
    AspVbs,
    Asterisk,
    AsteriskVoiceMail,
    Astro,
    Atlas,
    AutoHotKey,
    AutoIt,
    Automake,
    Ave,
    Awk,
    B,
    Bash,
    Basic,
    Bass,
    Bat,
    Bc,
    Bdf,
    BeanCount,
    Bib,
    Bicep,
    Bindzone,
    Bitbake,
    Blank,
    Blueprint,
    Bsdl,
    Bst,
    Btm,
    Bzl,
    Bzr,
    C,
    Cabal,
    CabalConfig,
    CabalProject,
    Cairo,
    Calendar,
    Capnp,
    Catalog,
    Cdc,
    Cdl,
    CdrdaoConf,
    Cdrtoc,
    Cf,
    CfEngine,
    Cfg,
    Ch,
    ChaiScript,
    Change,
    Changelog,
    Chaskell,
    Chatito,
    Chill,
    Chordpro,
    Cl,
    Clean,
    Clipper,
    Clojure,
    CMake,
    CMod,
    Cmusrc,
    Cobol,
    Coco,
    ConaryRecipe,
    Conf,
    Config,
    ConfIni,
    Context,
    Cook,
    Coq,
    Corn,
    Cpon,
    Cpp,
    Cqlang,
    Crm,
    Crontab,
    Crystal,
    Csc,
    Csdl,
    Csh,
    CSharp,
    Csp,
    Css,
    Csv,
    CTerm,
    Cucumber,
    Cuda,
    Cue,
    Cupl,
    Cuplsim,
    Cvs,
    Cvsrc,
    Cweb,
    Cynpp,
    D,
    Dart,
    DataScript,
    Dcd,
    Dcl,
    DebChangelog,
    DebControl,
    DebCopyright,
    DebSources,
    Def,
    DenyHosts,
    Dep3Patch,
    Desc,
    Desktop,
    Dhall,
    DictConf,
    DictdConf,
    Diff,
    DirColors,
    Diva,
    DnsMasq,
    DocBookSgml4,
    DocBookXml4,
    DocBookXml5,
    Dockerfile,
    DosBatch,
    DosIni,
    Dot,
    Dracula,
    Dsl,
    Dtd,
    DTrace,
    Dts,
    Dune,
    Dylan,
    DylanIntr,
    DylanLid,
    Ebnf,
    Ecd,
    Edif,
    EditorConfig,
    EElixir,
    Eiffel,
    Eighth,
    EJavaScript,
    Elf,
    ELinks,
    Elixir,
    Elm,
    ElmFilt,
    Elsa,
    Elvish,
    EPuppet,
    Erlang,
    ERuby,
    Esdl,
    Esmtprc,
    Esqlc,
    Esterel,
    Eterm,
    Euphoria3,
    Exim,
    Expect,
    Exports,
    Factor,
    Falcon,
    Fan,
    Fennel,
    FetchMail,
    Fgl,
    Firrtl,
    Fish,
    Foam,
    FocExec,
    Form,
    Forth,
    Fortran,
    FpcMake,
    FrameScript,
    FreeBasic,
    Fsh,
    FSharp,
    FsTab,
    Func,
    Fusion,
    Fvwm,
    Fvwm1,
    Fvwm2,
    Fvwm2M4,
    Gdb,
    Gdmo,
    GdResource,
    GdScript,
    GdShader,
    Gedcom,
    GemText,
    Gift,
    Git,
    GitAttributes,
    GitCommit,
    GitConfig,
    GitIgnore,
    GitOlite,
    GitRebase,
    GitSendEmail,
    Gkrellmrc,
    Gleam,
    Glsl,
    Gnash,
    GnuPlot,
    Go,
    GoMod,
    GoSum,
    GoWork,
    Gp,
    Gpg,
    Grads,
    GraphQl,
    Gretl,
    Groovy,
    Group,
    Grub,
    Gsp,
    Gtkrc,
    Gyp,
    Hack,
    Haml,
    Hamster,
    Handlebars,
    Hare,
    Haskell,
    Haste,
    HastePreProc,
    Hb,
    Hcl,
    Heex,
    Hercules,
    Hex,
    HexDump,
    Hgcommit,
    HJson,
    HlsPlaylist,
    Hog,
    Hollywood,
    Hoon,
    HostConf,
    HostsAccess,
    Html,
    HtmlDjango,
    HtmlM4,
    HtTest,
    I3Config,
    IBasic,
    IceMenu,
    Icon,
    Idl,
    Idlang,
    Indent,
    Inform,
    Initng,
    InitTab,
    InstallShield,
    IpFilter,
    Iss,
    Ist,
    J,
    Jal,
    Jam,
    Java,
    JavaCc,
    JavaScript,
    JavaScriptGlimmer,
    Jess,
    JGraph,
    Jovial,
    JProperties,
    Jq,
    Json,
    Json5,
    JsonC,
    JsonL,
    Jsonnet,
    Jsp,
    Jsx,
    Julia,
    KConfig,
    Kdl,
    Kivy,
    Kix,
    Kotlin,
    Krl,
    KScript,
    Ksh,
    Kwt,
    Lace,
    LambdaProlog,
    Larch,
    Latte,
    Ld,
    Ldif,
    Lean,
    Ledger,
    Less,
    Lex,
    Lftp,
    LHaskell,
    Libao,
    Lilo,
    Lilypond,
    Limits,
    Liquid,
    Lisp,
    Lite,
    LiteStep,
    Livebook,
    Llvm,
    LogCheck,
    LoginAccess,
    LoginDefs,
    Logtalk,
    Lotos,
    Lout,
    Lpc,
    Lsl,
    Lss,
    Lua,
    Luau,
    Lynx,
    Lyrics,
    M3Build,
    M3Quake,
    M4,
    Mail,
    MailAliases,
    MailCap,
    Make,
    Mallard,
    ManConf,
    Map,
    Maple,
    Markdown,
    Mason,
    Master,
    Matlab,
    Maxima,
    Mel,
    Mermaid,
    Meson,
    Messages,
    Mf,
    Mgl,
    Mgp,
    Mib,
    Mix,
    Mma,
    Mmix,
    Mmp,
    ModConf,
    Modsim3,
    Modula2,
    Modula3,
    Monk,
    Moo,
    MoonScript,
    Move,
    Mp,
    MPlayerConf,
    MpMetafun,
    Mrxvtrc,
    Msidl,
    MsMessages,
    Msql,
    Mupad,
    Murphi,
    Mush,
    Muttrc,
    MySql,
    N1ql,
    Named,
    Nanorc,
    Natural,
    Ncf,
    Neomuttrc,
    Netrc,
    Nginx,
    Nim,
    Ninja,
    Nix,
    Nqc,
    Nroff,
    Nsis,
    Nu,
    Obj,
    ObjC,
    ObjCpp,
    Obse,
    OCaml,
    OCamlInterface,
    Occam,
    Octave,
    Odin,
    OmniMark,
    Opam,
    OpenRoad,
    OpenScad,
    OpenVpn,
    Opl,
    Ora,
    Org,
    PamConf,
    PamEnv,
    Papp,
    Pascal,
    Passwd,
    Pbtxt,
    Pccts,
    Pcmk,
    Pdf,
    Pem,
    Perl,
    Pf,
    PfMain,
    Php,
    Pike,
    Pilrc,
    Pine,
    PInfo,
    PlainTex,
    Pli,
    Plm,
    Plp,
    Plsql,
    Po,
    Pod,
    PoeFilter,
    Poke,
    Pony,
    Postscr,
    Pov,
    PovIni,
    Ppd,
    Ppwiz,
    Prisma,
    Privoxy,
    Proc,
    ProcMail,
    Progress,
    Prolog,
    Promela,
    Proto,
    Protocols,
    Prql,
    Ps1,
    Ps1Xml,
    Psf,
    Psl,
    PtcapPrint,
    PtcapTerm,
    Pug,
    Puppet,
    Pyret,
    Pyrex,
    Python,
    Qb64,
    Ql,
    QmlDir,
    Quake,
    Quarto,
    R,
    Racc,
    Racket,
    Radiance,
    Raku,
    Raml,
    Rapid,
    RatPoison,
    Rbs,
    Rc,
    Rcs,
    Readline,
    Rebol,
    Redif,
    Registry,
    Rego,
    Remind,
    ReScript,
    Resolv,
    Reva,
    Rexx,
    RHelp,
    Rib,
    Rmd,
    Rnc,
    Rng,
    Rnoweb,
    Robot,
    Robots,
    Ron,
    RouterOs,
    Rpcgen,
    Rpgle,
    Rpl,
    Rrst,
    Rst,
    Rtf,
    Ruby,
    Rush,
    Rust,
    Sage,
    Samba,
    Sas,
    Sass,
    Sather,
    Sbt,
    Scala,
    Scdoc,
    Scheme,
    Scilab,
    Screen,
    Scss,
    Sd,
    Sdc,
    Sdl,
    Sed,
    Sensors,
    Services,
    SetSerial,
    Sexplib,
    Sgml,
    SgmlDecl,
    Sh,
    Sieve,
    Sil,
    Sile,
    Simula,
    Sinda,
    Sisu,
    Skill,
    Slang,
    Slice,
    SlpConf,
    SlpReg,
    SlpSpi,
    Slrnrc,
    Slrnsc,
    Sm,
    Smali,
    Smarty,
    Smcl,
    Smgllnx,
    Smil,
    Smith,
    Smithy,
    Sml,
    Snobol4,
    Solidity,
    Solution,
    Sparql,
    Spec,
    SpecMan,
    Spice,
    Spup,
    Spyce,
    Sql,
    Sqlj,
    Sqr,
    Squid,
    Squirrel,
    Srec,
    Srt,
    Ssa,
    SshConfig,
    SshdConfig,
    St,
    Starlark,
    Stata,
    Stp,
    Structurizr,
    Sudoers,
    Supercollider,
    Surface,
    Svelte,
    Svg,
    Svn,
    SwayConfig,
    Swift,
    SwiftGyb,
    Sysctl,
    Systemd,
    SystemVerilog,
    Tablegen,
    Tads,
    Tags,
    Tak,
    Tal,
    TaskData,
    TaskEdit,
    Tcl,
    Tcsh,
    Teal,
    Template,
    Teraterm,
    Terminfo,
    Terraform,
    TerraformVars,
    Tex,
    TexInfo,
    TexMF,
    Tf,
    Thrift,
    Tidy,
    Tilde,
    Tla,
    Tli,
    Tmux,
    Toml,
    Tpp,
    Trace32,
    Trasys,
    TreeSitterQuery,
    Treetop,
    Trustees,
    Tsalt,
    Tsscl,
    Tssgm,
    Tssop,
    Tsv,
    Tsx,
    Turtle,
    Tutor,
    Twig,
    TypeScript,
    TypeScriptGlimmer,
    Typst,
    Uc,
    UdevConf,
    UdevPerm,
    UdevRules,
    Uil,
    Ungrammar,
    UpdateDb,
    Upstart,
    UpstreamDat,
    UpstreamInstallLog,
    UpstreamLog,
    UrlShortcut,
    Usd,
    UsServerLog,
    Usw2KagtLog,
    V,
    Vala,
    Vb,
    Vdf,
    Vdmpp,
    Vdmrt,
    Vdmsl,
    Vera,
    Verilog,
    Verilogams,
    Vgrindefs,
    Vhdl,
    Vhs,
    Vim,
    VimHelp,
    VimInfo,
    Virata,
    Vmasm,
    Voscm,
    Vrml,
    Vroom,
    Vue,
    Wast,
    Wat,
    Wdl,
    Web,
    WebMacro,
    Wget,
    Wget2,
    WinBatch,
    Wit,
    Wml,
    Wsh,
    Wsml,
    WvDial,
    Xbl,
    XDefaults,
    XF86Conf,
    XF86Conf3,
    XF86Conf4,
    Xhtml,
    Xinetd,
    XMath,
    Xml,
    XModMap,
    Xpm,
    Xpm2,
    XQuery,
    Xs,
    Xsd,
    Xslt,
    Yacc,
    Yaml,
    Yang,
    Yuck,
    Z8a,
    Zig,
    Zimbu,
    ZimbuTempl,
    Zir,
    Zserio,
    Zsh,

}

A non-exhaustive list of text file types.

The type derives the following traits for convenience. For (de)serialization to/from strings, lowercase casing is used unless otherwise specified in the variants docs.

[strum::Display]: Display formatting and a .to_string() method

[strum::AsRefStr]: AsRef<str> impl

[strum::EnumString]: FromStr impl for turning strings into the corresponding variant

[strum::EnumVariantNames]: an associated VARIANTS constant containing the string names of all variants (requires [strum::VariantNames] to be in scope)

Clone, Copy

Debug

Hash

PartialEq, Eq

Default: the default is FileType::Text

serde serde::Serialize: serialize into a string

serde serde::Deserialize: deserialize from a string

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Text

A plain text file. This is the default variant. (De)serialized as text.

A2ps

(De)serialized as a2ps.

A65

(De)serialized as a65.

Aap

(De)serialized as aap.

Abap

(De)serialized as abap.

Abaqus

(De)serialized as abaqus.

Abc

(De)serialized as abc.

Abel

(De)serialized as abel.

Acedb

(De)serialized as acedb.

Ada

(De)serialized as ada.

Ahdl

(De)serialized as ahdl.

Aidl

(De)serialized as aidl.

AlsaConf

(De)serialized as alsaconf.

Aml

(De)serialized as aml.

Ampl

(De)serialized as ampl.

Ant

(De)serialized as ant.

Apache

(De)serialized as apache.

ApacheStyle

(De)serialized as apachestyle.

AppleScript

(De)serialized as applescript.

AptConf

(De)serialized as aptconf.

Arch

(De)serialized as arch.

Arduino

(De)serialized as arduino.

Art

(De)serialized as art.

AsciiDoc

(De)serialized as asciidoc.

Asm

(De)serialized as asm.

Asn

(De)serialized as asn.

AspPerl

(De)serialized as aspperl.

AspVbs

(De)serialized as aspvbs.

Asterisk

(De)serialized as asterisk.

AsteriskVoiceMail

(De)serialized as asteriskvm, not asteriskvoicemail.

Astro

(De)serialized as astro.

Atlas

(De)serialized as atlas.

AutoHotKey

(De)serialized as autohotkey.

AutoIt

(De)serialized as autoit.

Automake

(De)serialized as automake.

Ave

(De)serialized as ave.

Awk

(De)serialized as awk.

B

(De)serialized as b.

Bash

(De)serialized as bash.

Basic

(De)serialized as basic.

Bass

(De)serialized as bass.

Bat

(De)serialized as bat.

Bc

(De)serialized as bc.

Bdf

(De)serialized as bdf.

BeanCount

(De)serialized as beancount.

Bib

(De)serialized as bib.

Bicep

(De)serialized as bicep.

Bindzone

(De)serialized as bindzone.

Bitbake

(De)serialized as bitbake.

Blank

(De)serialized as blank.

Blueprint

(De)serialized as blueprint.

Bsdl

(De)serialized as bsdl.

Bst

(De)serialized as bst.

Btm

(De)serialized as btm.

Bzl

(De)serialized as bzl.

Bzr

(De)serialized as bzr.

C

(De)serialized as c.

Cabal

(De)serialized as cabal.

CabalConfig

(De)serialized as cabalconfig.

CabalProject

(De)serialized as cabalproject.

Cairo

(De)serialized as cairo.

Calendar

(De)serialized as calendar.

Capnp

(De)serialized as capnp.

Catalog

(De)serialized as catalog.

Cdc

(De)serialized as cdc.

Cdl

(De)serialized as cdl.

CdrdaoConf

(De)serialized as cdrdaoconf.

Cdrtoc

(De)serialized as cdrtoc.

Cf

(De)serialized as cf.

CfEngine

(De)serialized as cfengine.

Cfg

(De)serialized as cfg.

Ch

(De)serialized as ch.

ChaiScript

(De)serialized as chaiscript.

Change

(De)serialized as change.

Changelog

(De)serialized as changelog.

Chaskell

(De)serialized as chaskell.

Chatito

(De)serialized as chatito.

Chill

(De)serialized as chill.

Chordpro

(De)serialized as chordpro.

Cl

(De)serialized as cl.

Clean

(De)serialized as clean.

Clipper

(De)serialized as clipper.

Clojure

(De)serialized as clojure.

CMake

(De)serialized as cmake.

CMod

(De)serialized as cmod.

Cmusrc

(De)serialized as cmusrc.

Cobol

(De)serialized as cobol.

Coco

(De)serialized as coco.

ConaryRecipe

(De)serialized as conaryrecipe.

Conf

(De)serialized as conf.

Config

(De)serialized as config.

ConfIni

(De)serialized as confini.

Context

(De)serialized as context.

Cook

(De)serialized as cook.

Coq

(De)serialized as coq.

Corn

(De)serialized as corn.

Cpon

(De)serialized as cpon.

Cpp

(De)serialized as cpp.

Cqlang

(De)serialized as cqlang.

Crm

(De)serialized as crm.

Crontab

(De)serialized as crontab.

Crystal

(De)serialized as crystal.

Csc

(De)serialized as csc.

Csdl

(De)serialized as csdl.

Csh

(De)serialized as csh.

CSharp

(De)serialized as csharp.

Csp

(De)serialized as csp.

Css

(De)serialized as css.

Csv

(De)serialized as csv.

CTerm

(De)serialized as cterm.

Cucumber

(De)serialized as cucumber.

Cuda

(De)serialized as cuda.

Cue

(De)serialized as cue.

Cupl

(De)serialized as cupl.

Cuplsim

(De)serialized as cuplsim.

Cvs

(De)serialized as cvs.

Cvsrc

(De)serialized as cvsrc.

Cweb

(De)serialized as cweb.

Cynpp

(De)serialized as cynpp.

D

(De)serialized as d.

Dart

(De)serialized as dart.

DataScript

(De)serialized as datascript.

Dcd

(De)serialized as dcd.

Dcl

(De)serialized as dcl.

DebChangelog

(De)serialized as debchangelog.

DebControl

(De)serialized as debcontrol.

DebCopyright

(De)serialized as debcopyright.

DebSources

(De)serialized as debsources.

Def

(De)serialized as def.

DenyHosts

(De)serialized as denyhosts.

Dep3Patch

(De)serialized as dep3patch.

Desc

(De)serialized as desc.

Desktop

(De)serialized as desktop.

Dhall

(De)serialized as dhall.

DictConf

(De)serialized as dictconf.

DictdConf

(De)serialized as dictdconf.

Diff

(De)serialized as diff.

DirColors

(De)serialized as dircolors.

Diva

(De)serialized as diva.

DnsMasq

(De)serialized as dnsmasq.

DocBookSgml4

(De)serialized as docbk-sgml-4, not docbooksgml4.

DocBookXml4

(De)serialized as docbk-xml-4, not docbookxml4.

DocBookXml5

(De)serialized as docbk-xml-5, not docbookxml5.

Dockerfile

(De)serialized as dockerfile.

DosBatch

(De)serialized as dosbatch.

DosIni

(De)serialized as dosini.

Dot

(De)serialized as dot.

Dracula

(De)serialized as dracula.

Dsl

(De)serialized as dsl.

Dtd

(De)serialized as dtd.

DTrace

(De)serialized as dtrace.

Dts

(De)serialized as dts.

Dune

(De)serialized as dune.

Dylan

(De)serialized as dylan.

DylanIntr

(De)serialized as dylanintr.

DylanLid

(De)serialized as dylanlid.

Ebnf

(De)serialized as ebnf.

Ecd

(De)serialized as ecd.

Edif

(De)serialized as edif.

EditorConfig

(De)serialized as editorconfig.

EElixir

(De)serialized as eelixir.

Eiffel

(De)serialized as eiffel.

Eighth

(De)serialized as eighth.

EJavaScript

(De)serialized as ejavascript.

Elf

(De)serialized as elf.

ELinks

(De)serialized as elinks.

Elixir

(De)serialized as elixir.

Elm

(De)serialized as elm.

ElmFilt

(De)serialized as elmfilt.

Elsa

(De)serialized as elsa.

Elvish

(De)serialized as elvish.

EPuppet

(De)serialized as epuppet.

Erlang

(De)serialized as erlang.

ERuby

(De)serialized as eruby.

Esdl

(De)serialized as esdl.

Esmtprc

(De)serialized as esmtprc.

Esqlc

(De)serialized as esqlc.

Esterel

(De)serialized as esterel.

Eterm

(De)serialized as eterm.

Euphoria3

(De)serialized as euphoria3.

Exim

(De)serialized as exim.

Expect

(De)serialized as expect.

Exports

(De)serialized as exports.

Factor

(De)serialized as factor.

Falcon

(De)serialized as falcon.

Fan

(De)serialized as fan.

Fennel

(De)serialized as fennel.

FetchMail

(De)serialized as fetchmail.

Fgl

(De)serialized as fgl.

Firrtl

(De)serialized as firrtl.

Fish

(De)serialized as fish.

Foam

(De)serialized as foam.

FocExec

(De)serialized as focexec.

Form

(De)serialized as form.

Forth

(De)serialized as forth.

Fortran

(De)serialized as fortran.

FpcMake

(De)serialized as fpcmake.

FrameScript

(De)serialized as framescript.

FreeBasic

(De)serialized as freebasic.

Fsh

(De)serialized as fsh.

FSharp

(De)serialized as fsharp.

FsTab

(De)serialized as fstab.

Func

(De)serialized as func.

Fusion

(De)serialized as fusion.

Fvwm

(De)serialized as fvwm.

Fvwm1

(De)serialized as fvwm-1, not fvwm1.

Fvwm2

(De)serialized as fvwm-2, not fvwm2.

Fvwm2M4

(De)serialized as fvwm2m4.

Gdb

(De)serialized as gdb.

Gdmo

(De)serialized as gdmo.

GdResource

(De)serialized as gdresource.

GdScript

(De)serialized as gdscript.

GdShader

(De)serialized as gdshader.

Gedcom

(De)serialized as gedcom.

GemText

(De)serialized as gemtext.

Gift

(De)serialized as gift.

Git

(De)serialized as git.

GitAttributes

(De)serialized as gitattributes.

GitCommit

(De)serialized as gitcommit.

GitConfig

(De)serialized as gitconfig.

GitIgnore

(De)serialized as gitignore.

GitOlite

(De)serialized as gitolite.

GitRebase

(De)serialized as gitrebase.

GitSendEmail

(De)serialized as gitsendemail.

Gkrellmrc

(De)serialized as gkrellmrc.

Gleam

(De)serialized as gleam.

Glsl

(De)serialized as glsl.

Gnash

(De)serialized as gnash.

GnuPlot

(De)serialized as gnuplot.

Go

(De)serialized as go.

GoMod

(De)serialized as gomod.

GoSum

(De)serialized as gosum.

GoWork

(De)serialized as gowork.

Gp

(De)serialized as gp.

Gpg

(De)serialized as gpg.

Grads

(De)serialized as grads.

GraphQl

(De)serialized as graphql.

Gretl

(De)serialized as gretl.

Groovy

(De)serialized as groovy.

Group

(De)serialized as group.

Grub

(De)serialized as grub.

Gsp

(De)serialized as gsp.

Gtkrc

(De)serialized as gtkrc.

Gyp

(De)serialized as gyp.

Hack

(De)serialized as hack.

Haml

(De)serialized as haml.

Hamster

(De)serialized as hamster.

Handlebars

(De)serialized as handlebars.

Hare

(De)serialized as hare.

Haskell

(De)serialized as haskell.

Haste

(De)serialized as haste.

HastePreProc

(De)serialized as hastepreproc.

Hb

(De)serialized as hb.

Hcl

(De)serialized as hcl.

Heex

(De)serialized as heex.

Hercules

(De)serialized as hercules.

Hex

(De)serialized as hex.

HexDump

(De)serialized as hexdump.

Hgcommit

(De)serialized as hgcommit.

HJson

(De)serialized as hjson.

HlsPlaylist

(De)serialized as hlsplaylist.

Hog

(De)serialized as hog.

Hollywood

(De)serialized as hollywood.

Hoon

(De)serialized as hoon.

HostConf

(De)serialized as hostconf.

HostsAccess

(De)serialized as hostsaccess.

Html

(De)serialized as html.

HtmlDjango

(De)serialized as htmldjango.

HtmlM4

(De)serialized as htmlm4.

HtTest

(De)serialized as httest.

I3Config

(De)serialized as i3config.

IBasic

(De)serialized as ibasic.

IceMenu

(De)serialized as icemenu.

Icon

(De)serialized as icon.

Idl

(De)serialized as idl.

Idlang

(De)serialized as idlang.

Indent

(De)serialized as indent.

Inform

(De)serialized as inform.

Initng

(De)serialized as initng.

InitTab

(De)serialized as inittab.

InstallShield

(De)serialized as installshield.

IpFilter

(De)serialized as ipfilter.

Iss

(De)serialized as iss.

Ist

(De)serialized as ist.

J

(De)serialized as j.

Jal

(De)serialized as jal.

Jam

(De)serialized as jam.

Java

(De)serialized as java.

JavaCc

(De)serialized as javacc.

JavaScript

(De)serialized as javascript.

JavaScriptGlimmer

(De)serialized as javascript.glimmer, not javascriptglimmer.

Jess

(De)serialized as jess.

JGraph

(De)serialized as jgraph.

Jovial

(De)serialized as jovial.

JProperties

(De)serialized as jproperties.

Jq

(De)serialized as jq.

Json

(De)serialized as json.

Json5

(De)serialized as json5.

JsonC

(De)serialized as jsonc.

JsonL

(De)serialized as jsonl.

Jsonnet

(De)serialized as jsonnet.

Jsp

(De)serialized as jsp.

Jsx

(De)serialized as jsx.

Julia

(De)serialized as julia.

KConfig

(De)serialized as kconfig.

Kdl

(De)serialized as kdl.

Kivy

(De)serialized as kivy.

Kix

(De)serialized as kix.

Kotlin

(De)serialized as kotlin.

Krl

(De)serialized as krl.

KScript

(De)serialized as kscript.

Ksh

(De)serialized as ksh.

Kwt

(De)serialized as kwt.

Lace

(De)serialized as lace.

LambdaProlog

(De)serialized as lambdaprolog.

Larch

(De)serialized as larch.

Latte

(De)serialized as latte.

Ld

(De)serialized as ld.

Ldif

(De)serialized as ldif.

Lean

(De)serialized as lean.

Ledger

(De)serialized as ledger.

Less

(De)serialized as less.

Lex

(De)serialized as lex.

Lftp

(De)serialized as lftp.

LHaskell

(De)serialized as lhaskell.

Libao

(De)serialized as libao.

Lilo

(De)serialized as lilo.

Lilypond

(De)serialized as lilypond.

Limits

(De)serialized as limits.

Liquid

(De)serialized as liquid.

Lisp

(De)serialized as lisp.

Lite

(De)serialized as lite.

LiteStep

(De)serialized as litestep.

Livebook

(De)serialized as livebook.

Llvm

(De)serialized as llvm.

LogCheck

(De)serialized as logcheck.

LoginAccess

(De)serialized as loginaccess.

LoginDefs

(De)serialized as logindefs.

Logtalk

(De)serialized as logtalk.

Lotos

(De)serialized as lotos.

Lout

(De)serialized as lout.

Lpc

(De)serialized as lpc.

Lsl

(De)serialized as lsl.

Lss

(De)serialized as lss.

Lua

(De)serialized as lua.

Luau

(De)serialized as luau.

Lynx

(De)serialized as lynx.

Lyrics

(De)serialized as lyrics.

M3Build

(De)serialized as m3build.

M3Quake

(De)serialized as m3quake.

M4

(De)serialized as m4.

Mail

(De)serialized as mail.

MailAliases

(De)serialized as mailaliases.

MailCap

(De)serialized as mailcap.

Make

(De)serialized as make.

Mallard

(De)serialized as mallard.

ManConf

(De)serialized as manconf.

Map

(De)serialized as map.

Maple

(De)serialized as maple.

Markdown

(De)serialized as markdown.

Mason

(De)serialized as mason.

Master

(De)serialized as master.

Matlab

(De)serialized as matlab.

Maxima

(De)serialized as maxima.

Mel

(De)serialized as mel.

Mermaid

(De)serialized as mermaid.

Meson

(De)serialized as meson.

Messages

(De)serialized as messages.

Mf

(De)serialized as mf.

Mgl

(De)serialized as mgl.

Mgp

(De)serialized as mgp.

Mib

(De)serialized as mib.

Mix

(De)serialized as mix.

Mma

(De)serialized as mma.

Mmix

(De)serialized as mmix.

Mmp

(De)serialized as mmp.

ModConf

(De)serialized as modconf.

Modsim3

(De)serialized as modsim3.

Modula2

(De)serialized as modula2.

Modula3

(De)serialized as modula3.

Monk

(De)serialized as monk.

Moo

(De)serialized as moo.

MoonScript

(De)serialized as moonscript.

Move

(De)serialized as move.

Mp

(De)serialized as mp.

MPlayerConf

(De)serialized as mplayerconf.

MpMetafun

(De)serialized as mp-metafun, not mpmetafun.

Mrxvtrc

(De)serialized as mrxvtrc.

Msidl

(De)serialized as msidl.

MsMessages

(De)serialized as msmessages.

Msql

(De)serialized as msql.

Mupad

(De)serialized as mupad.

Murphi

(De)serialized as murphi.

Mush

(De)serialized as mush.

Muttrc

(De)serialized as muttrc.

MySql

(De)serialized as mysql.

N1ql

(De)serialized as n1ql.

Named

(De)serialized as named.

Nanorc

(De)serialized as nanorc.

Natural

(De)serialized as natural.

Ncf

(De)serialized as ncf.

Neomuttrc

(De)serialized as neomuttrc.

Netrc

(De)serialized as netrc.

Nginx

(De)serialized as nginx.

Nim

(De)serialized as nim.

Ninja

(De)serialized as ninja.

Nix

(De)serialized as nix.

Nqc

(De)serialized as nqc.

Nroff

(De)serialized as nroff.

Nsis

(De)serialized as nsis.

Nu

(De)serialized as nu.

Obj

(De)serialized as obj.

ObjC

(De)serialized as objc.

ObjCpp

(De)serialized as objcpp.

Obse

(De)serialized as obse.

OCaml

(De)serialized as ocaml.

OCamlInterface

(De)serialized as ocamlinterface.

Occam

(De)serialized as occam.

Octave

(De)serialized as octave.

Odin

(De)serialized as odin.

OmniMark

(De)serialized as omnimark.

Opam

(De)serialized as opam.

OpenRoad

(De)serialized as openroad.

OpenScad

(De)serialized as openscad.

OpenVpn

(De)serialized as openvpn.

Opl

(De)serialized as opl.

Ora

(De)serialized as ora.

Org

(De)serialized as org.

PamConf

(De)serialized as pamconf.

PamEnv

(De)serialized as pamenv.

Papp

(De)serialized as papp.

Pascal

(De)serialized as pascal.

Passwd

(De)serialized as passwd.

Pbtxt

(De)serialized as pbtxt.

Pccts

(De)serialized as pccts.

Pcmk

(De)serialized as pcmk.

Pdf

(De)serialized as pdf.

Pem

(De)serialized as pem.

Perl

(De)serialized as perl.

Pf

(De)serialized as pf.

PfMain

(De)serialized as pfmain.

Php

(De)serialized as php.

Pike

(De)serialized as pike.

Pilrc

(De)serialized as pilrc.

Pine

(De)serialized as pine.

PInfo

(De)serialized as pinfo.

PlainTex

(De)serialized as plaintex.

Pli

(De)serialized as pli.

Plm

(De)serialized as plm.

Plp

(De)serialized as plp.

Plsql

(De)serialized as plsql.

Po

(De)serialized as po.

Pod

(De)serialized as pod.

PoeFilter

(De)serialized as poefilter.

Poke

(De)serialized as poke.

Pony

(De)serialized as pony.

Postscr

(De)serialized as postscr.

Pov

(De)serialized as pov.

PovIni

(De)serialized as povini.

Ppd

(De)serialized as ppd.

Ppwiz

(De)serialized as ppwiz.

Prisma

(De)serialized as prisma.

Privoxy

(De)serialized as privoxy.

Proc

(De)serialized as proc.

ProcMail

(De)serialized as procmail.

Progress

(De)serialized as progress.

Prolog

(De)serialized as prolog.

Promela

(De)serialized as promela.

Proto

(De)serialized as proto.

Protocols

(De)serialized as protocols.

Prql

(De)serialized as prql.

Ps1

(De)serialized as ps1.

Ps1Xml

(De)serialized as ps1xml.

Psf

(De)serialized as psf.

Psl

(De)serialized as psl.

PtcapPrint

(De)serialized as ptcap-print, not ptcapprint.

PtcapTerm

(De)serialized as ptcap-term, not ptcapterm.

Pug

(De)serialized as pug.

Puppet

(De)serialized as puppet.

Pyret

(De)serialized as pyret.

Pyrex

(De)serialized as pyrex.

Python

(De)serialized as python.

Qb64

(De)serialized as qb64.

Ql

(De)serialized as ql.

QmlDir

(De)serialized as qmldir.

Quake

(De)serialized as quake.

Quarto

(De)serialized as quarto.

R

(De)serialized as r.

Racc

(De)serialized as racc.

Racket

(De)serialized as racket.

Radiance

(De)serialized as radiance.

Raku

(De)serialized as raku.

Raml

(De)serialized as raml.

Rapid

(De)serialized as rapid.

RatPoison

(De)serialized as ratpoison.

Rbs

(De)serialized as rbs.

Rc

(De)serialized as rc.

Rcs

(De)serialized as rcs.

Readline

(De)serialized as readline.

Rebol

(De)serialized as rebol.

Redif

(De)serialized as redif.

Registry

(De)serialized as registry.

Rego

(De)serialized as rego.

Remind

(De)serialized as remind.

ReScript

(De)serialized as rescript.

Resolv

(De)serialized as resolv.

Reva

(De)serialized as reva.

Rexx

(De)serialized as rexx.

RHelp

(De)serialized as rhelp.

Rib

(De)serialized as rib.

Rmd

(De)serialized as rmd.

Rnc

(De)serialized as rnc.

Rng

(De)serialized as rng.

Rnoweb

(De)serialized as rnoweb.

Robot

(De)serialized as robot.

Robots

(De)serialized as robots.

Ron

(De)serialized as ron.

RouterOs

(De)serialized as routeros.

Rpcgen

(De)serialized as rpcgen.

Rpgle

(De)serialized as rpgle.

Rpl

(De)serialized as rpl.

Rrst

(De)serialized as rrst.

Rst

(De)serialized as rst.

Rtf

(De)serialized as rtf.

Ruby

(De)serialized as ruby.

Rush

(De)serialized as rush.

Rust

(De)serialized as rust.

Sage

(De)serialized as sage.

Samba

(De)serialized as samba.

Sas

(De)serialized as sas.

Sass

(De)serialized as sass.

Sather

(De)serialized as sather.

Sbt

(De)serialized as sbt.

Scala

(De)serialized as scala.

Scdoc

(De)serialized as scdoc.

Scheme

(De)serialized as scheme.

Scilab

(De)serialized as scilab.

Screen

(De)serialized as screen.

Scss

(De)serialized as scss.

Sd

(De)serialized as sd.

Sdc

(De)serialized as sdc.

Sdl

(De)serialized as sdl.

Sed

(De)serialized as sed.

Sensors

(De)serialized as sensors.

Services

(De)serialized as services.

SetSerial

(De)serialized as setserial.

Sexplib

(De)serialized as sexplib.

Sgml

(De)serialized as sgml.

SgmlDecl

(De)serialized as sgmldecl.

Sh

(De)serialized as sh.

Sieve

(De)serialized as sieve.

Sil

(De)serialized as sil.

Sile

(De)serialized as sile.

Simula

(De)serialized as simula.

Sinda

(De)serialized as sinda.

Sisu

(De)serialized as sisu.

Skill

(De)serialized as skill.

Slang

(De)serialized as slang.

Slice

(De)serialized as slice.

SlpConf

(De)serialized as slpconf.

SlpReg

(De)serialized as slpreg.

SlpSpi

(De)serialized as slpspi.

Slrnrc

(De)serialized as slrnrc.

Slrnsc

(De)serialized as slrnsc.

Sm

(De)serialized as sm.

Smali

(De)serialized as smali.

Smarty

(De)serialized as smarty.

Smcl

(De)serialized as smcl.

Smgllnx

(De)serialized as smgllnx.

Smil

(De)serialized as smil.

Smith

(De)serialized as smith.

Smithy

(De)serialized as smithy.

Sml

(De)serialized as sml.

Snobol4

(De)serialized as snobol4.

Solidity

(De)serialized as solidity.

Solution

(De)serialized as solution.

Sparql

(De)serialized as sparql.

Spec

(De)serialized as spec.

SpecMan

(De)serialized as specman.

Spice

(De)serialized as spice.

Spup

(De)serialized as spup.

Spyce

(De)serialized as spyce.

Sql

(De)serialized as sql.

Sqlj

(De)serialized as sqlj.

Sqr

(De)serialized as sqr.

Squid

(De)serialized as squid.

Squirrel

(De)serialized as squirrel.

Srec

(De)serialized as srec.

Srt

(De)serialized as srt.

Ssa

(De)serialized as ssa.

SshConfig

(De)serialized as sshconfig.

SshdConfig

(De)serialized as sshdconfig.

St

(De)serialized as st.

Starlark

(De)serialized as starlark.

Stata

(De)serialized as stata.

Stp

(De)serialized as stp.

Structurizr

(De)serialized as structurizr.

Sudoers

(De)serialized as sudoers.

Supercollider

(De)serialized as supercollider.

Surface

(De)serialized as surface.

Svelte

(De)serialized as svelte.

Svg

(De)serialized as svg.

Svn

(De)serialized as svn.

SwayConfig

(De)serialized as swayconfig.

Swift

(De)serialized as swift.

SwiftGyb

(De)serialized as swiftgyb.

Sysctl

(De)serialized as sysctl.

Systemd

(De)serialized as systemd.

SystemVerilog

(De)serialized as systemverilog.

Tablegen

(De)serialized as tablegen.

Tads

(De)serialized as tads.

Tags

(De)serialized as tags.

Tak

(De)serialized as tak.

Tal

(De)serialized as tal.

TaskData

(De)serialized as taskdata.

TaskEdit

(De)serialized as taskedit.

Tcl

(De)serialized as tcl.

Tcsh

(De)serialized as tcsh.

Teal

(De)serialized as teal.

Template

(De)serialized as template.

Teraterm

(De)serialized as teraterm.

Terminfo

(De)serialized as terminfo.

Terraform

(De)serialized as terraform.

TerraformVars

(De)serialized as terraform-vars, not terraformvars.

Tex

(De)serialized as tex.

TexInfo

(De)serialized as texinfo.

TexMF

(De)serialized as texmf.

Tf

(De)serialized as tf.

Thrift

(De)serialized as thrift.

Tidy

(De)serialized as tidy.

Tilde

(De)serialized as tilde.

Tla

(De)serialized as tla.

Tli

(De)serialized as tli.

Tmux

(De)serialized as tmux.

Toml

(De)serialized as toml.

Tpp

(De)serialized as tpp.

Trace32

(De)serialized as trace32.

Trasys

(De)serialized as trasys.

TreeSitterQuery

(De)serialized as query, not treesitterquery.

Treetop

(De)serialized as treetop.

Trustees

(De)serialized as trustees.

Tsalt

(De)serialized as tsalt.

Tsscl

(De)serialized as tsscl.

Tssgm

(De)serialized as tssgm.

Tssop

(De)serialized as tssop.

Tsv

(De)serialized as tsv.

Tsx

(De)serialized as tsx.

Turtle

(De)serialized as turtle.

Tutor

(De)serialized as tutor.

Twig

(De)serialized as twig.

TypeScript

(De)serialized as typescript.

TypeScriptGlimmer

(De)serialized as typescript.glimmer, not typescriptglimmer.

Typst

(De)serialized as typst.

Uc

(De)serialized as uc.

UdevConf

(De)serialized as udevconf.

UdevPerm

(De)serialized as udevperm.

UdevRules

(De)serialized as udevrules.

Uil

(De)serialized as uil.

Ungrammar

(De)serialized as ungrammar.

UpdateDb

(De)serialized as updatedb.

Upstart

(De)serialized as upstart.

UpstreamDat

(De)serialized as upstreamdat.

UpstreamInstallLog

(De)serialized as upstreaminstalllog.

UpstreamLog

(De)serialized as upstreamlog.

UrlShortcut

(De)serialized as urlshortcut.

Usd

(De)serialized as usd.

UsServerLog

(De)serialized as usserverlog.

Usw2KagtLog

(De)serialized as usw2kagtlog.

V

(De)serialized as v.

Vala

(De)serialized as vala.

Vb

(De)serialized as vb.

Vdf

(De)serialized as vdf.

Vdmpp

(De)serialized as vdmpp.

Vdmrt

(De)serialized as vdmrt.

Vdmsl

(De)serialized as vdmsl.

Vera

(De)serialized as vera.

Verilog

(De)serialized as verilog.

Verilogams

(De)serialized as verilogams.

Vgrindefs

(De)serialized as vgrindefs.

Vhdl

(De)serialized as vhdl.

Vhs

(De)serialized as vhs.

Vim

(De)serialized as vim.

VimHelp

(De)serialized as help, not vimhelp.

VimInfo

(De)serialized as viminfo.

Virata

(De)serialized as virata.

Vmasm

(De)serialized as vmasm.

Voscm

(De)serialized as voscm.

Vrml

(De)serialized as vrml.

Vroom

(De)serialized as vroom.

Vue

(De)serialized as vue.

Wast

(De)serialized as wast.

Wat

(De)serialized as wat.

Wdl

(De)serialized as wdl.

Web

(De)serialized as web.

WebMacro

(De)serialized as webmacro.

Wget

(De)serialized as wget.

Wget2

(De)serialized as wget2.

WinBatch

(De)serialized as winbatch.

Wit

(De)serialized as wit.

Wml

(De)serialized as wml.

Wsh

(De)serialized as wsh.

Wsml

(De)serialized as wsml.

WvDial

(De)serialized as wvdial.

Xbl

(De)serialized as xbl.

XDefaults

(De)serialized as xdefaults.

XF86Conf

(De)serialized as xf86conf.

XF86Conf3

(De)serialized as xf86conf-3, not xf86conf3.

XF86Conf4

(De)serialized as xf86conf-4, not xf86conf4.

Xhtml

(De)serialized as xhtml.

Xinetd

(De)serialized as xinetd.

XMath

(De)serialized as xmath.

Xml

(De)serialized as xml.

XModMap

(De)serialized as xmodmap.

Xpm

(De)serialized as xpm.

Xpm2

(De)serialized as xpm2.

XQuery

(De)serialized as xquery.

Xs

(De)serialized as xs.

Xsd

(De)serialized as xsd.

Xslt

(De)serialized as xslt.

Yacc

(De)serialized as yacc.

Yaml

(De)serialized as yaml.

Yang

(De)serialized as yang.

Yuck

(De)serialized as yuck.

Z8a

(De)serialized as z8a.

Zig

(De)serialized as zig.

Zimbu

(De)serialized as zimbu.

ZimbuTempl

(De)serialized as zimbutempl.

Zir

(De)serialized as zir.

Zserio

(De)serialized as zserio.

Zsh

(De)serialized as zsh.