syntastica_core::language_set

Struct Language

pub struct Language(/* private fields */);

An opaque object that defines how to parse a particular language. The code for each Language is generated by the Tree-sitter CLI.

Implementations

impl Language

pub const unsafe fn from_raw(ptr: *const TSLanguage) -> Language

Reconstructs a Language from a raw pointer.

Safety

ptr must be non-null.

pub fn into_raw(self) -> *const TSLanguage

Consumes the Language, returning a raw pointer to the underlying C structure.

impl Language

pub fn new(builder: LanguageFn) -> Language

pub fn name(&self) -> Option<&'static str>

Get the name of this language. This returns None in older parsers.

pub fn version(&self) -> usize

๐Deprecated since 0.25.0: Use abi_version instead

Get the ABI version number that indicates which version of the Tree-sitter CLI that was used to generate this Language.

pub fn abi_version(&self) -> usize

Get the ABI version number that indicates which version of the Tree-sitter CLI that was used to generate this Language.

pub fn metadata(&self) -> Option<LanguageMetadata>

Get the metadata for this language. This information is generated by the CLI, and relies on the language author providing the correct metadata in the languageโs tree-sitter.json file.

See also [LanguageMetadata].

pub fn node_kind_count(&self) -> usize

Get the number of distinct node types in this language.

pub fn parse_state_count(&self) -> usize

Get the number of valid states in this language.

pub fn supertypes(&self) -> &[u16]

Get a list of all supertype symbols for the language.

pub fn subtypes_for_supertype(&self, supertype: u16) -> &[u16]

Get a list of all subtype symbols for a given supertype symbol.

pub fn node_kind_for_id(&self, id: u16) -> Option<&'static str>

Get the name of the node kind for the given numerical id.

pub fn id_for_node_kind(&self, kind: &str, named: bool) -> u16

Get the numeric id for the given node kind.

pub fn node_kind_is_named(&self, id: u16) -> bool

Check if the node type for the given numerical id is named (as opposed to an anonymous node type).

pub fn node_kind_is_visible(&self, id: u16) -> bool

Check if the node type for the given numerical id is visible (as opposed to a hidden node type).

pub fn node_kind_is_supertype(&self, id: u16) -> bool

Check if the node type for the given numerical id is a supertype.

pub fn field_count(&self) -> usize

Get the number of distinct field names in this language.

pub fn field_name_for_id(&self, field_id: u16) -> Option<&'static str>

Get the field name for the given numerical id.

pub fn field_id_for_name( &self, field_name: impl AsRef<[u8]>, ) -> Option<NonZero<u16>>

Get the numerical id for the given field name.

pub fn next_state(&self, state: u16, id: u16) -> u16

Get the next parse state. Combine this with lookahead_iterator to generate completion suggestions or valid symbols in error nodes.

Example:

let state = language.next_state(node.parse_state(), node.grammar_id());

pub fn lookahead_iterator(&self, state: u16) -> Option<LookaheadIterator>

Create a new lookahead iterator for this language and parse state.

This returns None if state is invalid for this language.

Iterating [LookaheadIterator] will yield valid symbols in the given parse state. Newly created lookahead iterators will return the ERROR symbol from [LookaheadIterator::current_symbol].

Lookahead iterators can be useful to generate suggestions and improve syntax error diagnostics. To get symbols valid in an ERROR node, use the lookahead iterator on its first leaf node state. For MISSING nodes, a lookahead iterator created on the previous non-extra leaf node may be appropriate.