Trait syntastica_core::provider::LanguageProvider

pub trait LanguageProvider {
    // Required methods
    fn get_parsers(&self) -> Result<Parsers, Error>;
    fn get_queries(&self) -> Result<Queries<'_>, Error>;
    fn for_extension<'a>(&self, file_extension: &'a str) -> Option<Cow<'a, str>>;

    // Provided methods
    fn for_injection<'a>(&self, name: &'a str) -> Option<Cow<'a, str>> { ... }
    fn get_languages(&self) -> Result<Languages, Error> { ... }
}

Describes a type that is able to provide tree-sitter parsers and queries.

All three syntastica parser collections provide a type called LanguageProviderImpl which implements this trait. See the project overview for more information on that.

A LanguageProvider has two different uses:

Providing tree-sitter parsers and queries (see get_parsers, get_queries, and get_languages)

Resolving a language name based on a file extension or injection name (see for_extension and for_injection)

Required Methods

fn get_parsers(&self) -> Result<Parsers, Error>

Provide parsers for all supported (or selected) languages this provider has to offer.

For every entry in the returned map, the map returned by get_queries must contain an entry with the same key.

See Parsers for information on the return type.

fn get_queries(&self) -> Result<Queries<'_>, Error>

Provide queries for all supported (or selected) languages this provider has to offer.

For every entry in the map returned by get_parsers, the map returned here must contain an entry with the same key.

See Queries for information on the return type.

fn for_extension<'a>(&self, file_extension: &'a str) -> Option<Cow<'a, str>>

Find a language based on the given file extension.

If the provider supports the given file extension, then the name of the language that file extension belongs to should be returned. This name must be a key in the map returned by get_languages. If the file extension is not supported, None should be returned.

Provided Methods

fn for_injection<'a>(&self, name: &'a str) -> Option<Cow<'a, str>>

Find a language for an injection.

The passed name could be any string that somehow identifies a language. This very much depends on the source of the injection. For example, in fenced code blocks in markdown, the text after the opening ``` is passed to this function.

Note that this function does not get called, if the name is a valid key in the map returned by get_languages.

If the provider supports the requested language, then the name of the language that should be used for the injection should be returned. This name must be a key in the map returned by get_languages. If no matching language is found, None should be returned.

The default implementation forwards to for_extension.

fn get_languages(&self) -> Result<Languages, Error>

Provide all supported (or selected) languages this provider has to offer.

See Languages for information on the return type.

The default implementation calls get_parsers and get_queries, and merges the results.